# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EarnPearls API",
    "description": "Version 1 API for authentication, surveys, wallet, withdrawals, and administration.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ep_session",
        "description": "Opaque, revocable HttpOnly session cookie."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Session-bound token required for authenticated state changes."
      },
      "providerSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Provider-Signature",
        "description": "Provider-adapter-specific webhook signature."
      }
    },
...
  },
  "paths": {
...
    "/wallet/": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "pending",
                    "validated",
                    "mature",
                    "withdrawable",
                    "reserved",
                    "paid",
                    "rejected",
                    "reversed",
                    "totalEarnings",
                    "conversion"
                  ],
                  "properties": {
                    "pending": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "validated": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "mature": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "withdrawable": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "reserved": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "paid": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "rejected": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "reversed": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "totalEarnings": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "conversion": {
                      "type": "object",
                      "required": [
                        "pointsPerUsd",
                        "sourceCurrency",
                        "localCurrencyEstimate"
                      ],
                      "properties": {
                        "pointsPerUsd": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "sourceCurrency": {
                          "type": "string",
                          "enum": [
                            "USD"
                          ]
                        },
                        "localCurrencyEstimate": {
                          "type": "null"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/surveys/": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "title",
                      "reward",
                      "estimatedMinutes",
                      "difficulty",
                      "category",
                      "deviceCompatibility",
                      "countryEligible",
                      "available"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "reward": {
                        "type": "object",
                        "required": [
                          "points",
                          "usdMicros",
                          "usd"
                        ],
                        "properties": {
                          "points": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usdMicros": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usd": {
                            "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                            "type": "string"
                          }
                        }
                      },
                      "estimatedMinutes": {
                        "anyOf": [
                          {
                            "minimum": 1,
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "difficulty": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "category": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "deviceCompatibility": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "countryEligible": {
                        "type": "boolean"
                      },
                      "available": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/withdrawals/": {
      "get": {
        "tags": [
          "Withdrawals"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "methodCode",
                      "amount",
                      "fee",
                      "status",
                      "destinationMasked",
                      "requestedAt",
                      "updatedAt",
                      "processedAt",
                      "rejectionReason"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "methodCode": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "object",
                        "required": [
                          "points",
                          "usdMicros",
                          "usd"
                        ],
                        "properties": {
                          "points": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usdMicros": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usd": {
                            "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                            "type": "string"
                          }
                        }
                      },
                      "fee": {
                        "type": "object",
                        "required": [
                          "points",
                          "usdMicros",
                          "usd"
                        ],
                        "properties": {
                          "points": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usdMicros": {
                            "pattern": "^-?[0-9]+$",
                            "type": "string"
                          },
                          "usd": {
                            "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                            "type": "string"
                          }
                        }
                      },
                      "status": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "requested"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "under_review"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "approved"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "processing"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "paid"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "rejected"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "cancelled"
                            ]
                          }
                        ]
                      },
                      "destinationMasked": {
                        "type": "string"
                      },
                      "requestedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "processedAt": {
                        "anyOf": [
                          {
                            "format": "date-time",
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "rejectionReason": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Withdrawals"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "methodCode",
                  "points",
                  "destination"
                ],
                "properties": {
                  "methodCode": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  },
                  "points": {
                    "pattern": "^[1-9][0-9]*$",
                    "type": "string"
                  },
                  "destination": {
                    "minLength": 3,
                    "maxLength": 500,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 16,
              "maxLength": 200,
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "methodCode",
                    "amount",
                    "fee",
                    "status",
                    "destinationMasked",
                    "requestedAt",
                    "updatedAt",
                    "processedAt",
                    "rejectionReason"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "methodCode": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "fee": {
                      "type": "object",
                      "required": [
                        "points",
                        "usdMicros",
                        "usd"
                      ],
                      "properties": {
                        "points": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usdMicros": {
                          "pattern": "^-?[0-9]+$",
                          "type": "string"
                        },
                        "usd": {
                          "pattern": "^-?[0-9]+\\.[0-9]{6}$",
                          "type": "string"
                        }
                      }
                    },
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "requested"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "under_review"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "approved"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "processing"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "paid"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "rejected"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "cancelled"
                          ]
                        }
                      ]
                    },
                    "destinationMasked": {
                      "type": "string"
                    },
                    "requestedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "processedAt": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rejectionReason": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/withdrawals": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "schema": {
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "userId",
                      "userEmail",
                      "methodCode",
                      "points",
                      "status",
                      "destinationMasked",
                      "requestedAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "userId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "userEmail": {
                        "format": "email",
                        "type": "string"
                      },
                      "methodCode": {
                        "type": "string"
                      },
                      "points": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "destinationMasked": {
                        "type": "string"
                      },
                      "requestedAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/v1",
      "description": "Current origin"
    }
  ],
...
}